requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bootstrap4==0.1.0
Django==4.1.6
django-bootstrap4==22.3
django-tinymce==3.5.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
